requirements.txt
requests>=2.28.0
pytest>=7.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
